requirements.txt
Flask
Flask-SQLAlchemy  
python-dotenv     
boto3             
SpeechRecognition 